move to separate conftest
