+++ src/mrpro/operators/Functional.py
@@ -1,7 +1,5 @@
 """Base Class Functional."""
 
-from __future__ import annotations
-
 import math
 from abc import ABC, abstractmethod
 from collections.abc import Sequence
