+++ tests/operators/functionals/test_functionals.py
@@ -1,99 +1,24 @@
-import inspect
-from collections.abc import Callable, Sequence
 from copy import deepcopy
-from dataclasses import dataclass
 from typing import Literal, TypedDict
 
 import pytest
 import torch
-from mrpro.operators import Functional, ProximableFunctional, functionals
+from mrpro.operators.Functional import ElementaryFunctional, ElementaryProximableFunctional
 
 from tests import RandomGenerator
-
-
-@dataclass
-class FunctionalTestCase:
-    """A testcase for proximable functionals used in test parametrization.
-
-    See functional_test_cases.
-
-    """
-
-    functional: ProximableFunctional
-    x_dtype: torch.dtype
-    x_shape: torch.Size
-    rng: RandomGenerator
-    sigma: float | torch.Tensor
-
-    def rand_x(self) -> torch.Tensor:
-        """Generate random x"""
-        low = 0 if self.x_dtype.is_complex else -2
-        return self.rng.rand_tensor(self.x_shape, low=low, high=2.0, dtype=self.x_dtype)
-
-    @property
-    def result_dtype(self):
-        """Expected result dtype"""
-        return torch.promote_types(self.x_dtype, torch.result_type(self.functional.target, self.functional.weight))
-
-
-FUNCTIONALS: list[type[Functional]] = [
-    x[1] for x in inspect.getmembers(functionals, lambda x: inspect.isclass(x) and issubclass(x, Functional))
-]
-PROXIMABLE_FUNCTIONALS: list[type[ProximableFunctional]] = [
-    x for x in FUNCTIONALS if issubclass(x, ProximableFunctional)
-]
-
-
-def functional_test_cases(func: Callable[[FunctionalTestCase], None]) -> None:
-    """Decorator combining multiple parameterizations for test cases for all proximable functionals."""
-
-    @pytest.mark.parametrize('shape', [[1, 2, 3]])
-    @pytest.mark.parametrize('dtype_name', ['float32', 'complex64'])
-    @pytest.mark.parametrize('weight', ['scalar_weight', 'tensor_weight', 'complex_weight'])
-    @pytest.mark.parametrize('target', ['no_target', 'random_target'])
-    @pytest.mark.parametrize('dim', [None])
-    @pytest.mark.parametrize('divide_by_n', [True, False])
-    @pytest.mark.parametrize('functional', FUNCTIONALS)
-    def wrapper(
-        functional: type[ProximableFunctional],
-        shape: torch.Size,
-        dtype_name: Literal['float32', 'float64', 'complex64', 'complex128'],
-        weight: Literal['scalar_weight', 'tensor_weight', 'binary_weight', 'complex_weight'],
-        target: Literal['no_target', 'random_target', 'zero_target'],
-        dim: None | Sequence[int],
-        divide_by_n: bool,
-    ):
-        dtype = getattr(torch, dtype_name)
-        rng = RandomGenerator(13)
-        sigma = rng.rand_tensor((1,), low=0.0, high=1.0, dtype=dtype.to_real())
-        match weight:
-            case 'scalar_weight':
-                weight_value: float | torch.Tensor = rng.float64(low=-2, high=2)
-            case 'binary_weight':
-                weight_value = rng.rand_tensor(shape, low=0.0, high=1.0, dtype=dtype.to_real()).round()
-            case 'complex_weight':
-                weight_value = rng.rand_tensor(shape, low=0.0, high=1.0, dtype=dtype.to_complex())
-            case 'tensor_weight':
-                weight_value = rng.rand_tensor(shape, low=-1.0, high=1.0, dtype=dtype.to_real())
-        match target:
-            case 'zero_target':
-                target_value: torch.Tensor | None = torch.zeros(shape, dtype=dtype)
-            case 'no_target':
-                target_value = None
-            case 'random_target':
-                target_value = rng.rand_tensor(shape, low=0 if dtype.is_complex else -1, high=1.0, dtype=dtype)
-        f = functional(divide_by_n=divide_by_n, weight=weight_value, target=target_value, dim=dim)
-        case = FunctionalTestCase(functional=f, x_shape=shape, x_dtype=dtype, rng=rng, sigma=sigma)
-        func(case)
-
-    return wrapper
+from tests.operators.functionals.conftest import (
+    FUNCTIONALS,
+    PROXIMABLE_FUNCTIONALS,
+    FunctionalTestCase,
+    functional_test_cases,
+)
 
 
 @pytest.mark.parametrize('shape', [[1, 2, 3], ()])
 @pytest.mark.parametrize('dim', [None, -2])
 @pytest.mark.parametrize('keepdim', [True, False])
 @pytest.mark.parametrize('functional', FUNCTIONALS)
-def test_functional_shape(functional: type[Functional], shape: torch.Size, dim: None | int, keepdim: bool):
+def test_functional_shape(functional: type[ElementaryFunctional], shape: torch.Size, dim: None | int, keepdim: bool):
     """Test the shape returned by the forward of the functional."""
     f = functional(dim=dim, keepdim=keepdim)
     x = torch.ones(shape)
@@ -128,7 +53,7 @@ def test_functional_shape(functional: type[Functional], shape: torch.Size, dim: 
 @pytest.mark.parametrize('keepdim', [True, False])
 @pytest.mark.parametrize('functional', FUNCTIONALS)
 def test_functional_divide_by_n(
-    functional: type[Functional], shape: torch.Size, dim: None | tuple[int, ...], keepdim: bool
+    functional: type[ElementaryFunctional], shape: torch.Size, dim: None | tuple[int, ...], keepdim: bool
 ):
     """Test if divide_by_n scales by number of elements indexed by dim."""
 
@@ -251,7 +176,7 @@ def test_functional_prox_zero_weight(case: FunctionalTestCase):
 @pytest.mark.parametrize('function', ['prox', 'prox_convex_conj'])
 @pytest.mark.parametrize('functional', PROXIMABLE_FUNCTIONALS)
 def test_functional_prox_zero_sigma(
-    functional: type[ProximableFunctional], function: Literal['prox', 'prox_convex_conj']
+    functional: type[ElementaryProximableFunctional], function: Literal['prox', 'prox_convex_conj']
 ):
     """Test prox with sigma=0"""
     p = getattr(functional(), function)
@@ -265,7 +190,7 @@ def test_functional_prox_zero_sigma(
 @pytest.mark.parametrize('function', ['prox', 'prox_convex_conj'])
 @pytest.mark.parametrize('functional', PROXIMABLE_FUNCTIONALS)
 def test_functional_negative_sigma(
-    functional: type[ProximableFunctional], function: Literal['prox', 'prox_convex_conj']
+    functional: type[ElementaryProximableFunctional], function: Literal['prox', 'prox_convex_conj']
 ):
     """Test if error is raised for negative or complex sigma"""
 
@@ -302,7 +227,7 @@ class NumericCase(TypedDict):
     }
     """
 
-    functional: type[ProximableFunctional]
+    functional: type[ElementaryProximableFunctional]
     x: torch.Tensor
     weight: complex | torch.Tensor
     target: torch.Tensor
@@ -324,7 +249,7 @@ def test_functional_values(case_name: str):
 
 
 def _test_functional_values(
-    functional: type[ProximableFunctional],
+    functional: type[ElementaryProximableFunctional],
     x: torch.Tensor,
     weight: complex | torch.Tensor,
     target: torch.Tensor,
@@ -352,7 +277,9 @@ def test_functional_has_testcase(functional):
 @pytest.mark.cuda()
 @pytest.mark.parametrize('functional', FUNCTIONALS)
 @pytest.mark.parametrize('parameters', ['scalar', 'none', 'tensor'])
-def test_functional_cuda_forward(functional: type[Functional], parameters: Literal['scalar', 'none', 'tensor']):
+def test_functional_cuda_forward(
+    functional: type[ElementaryFunctional], parameters: Literal['scalar', 'none', 'tensor']
+):
     """Test if the forward pass works on the GPU."""
     match parameters:
         case 'scalar':
@@ -370,7 +297,9 @@ def test_functional_cuda_forward(functional: type[Functional], parameters: Liter
 @pytest.mark.cuda()
 @pytest.mark.parametrize('parameters', ['scalar', 'none', 'tensor'])
 @pytest.mark.parametrize('functional', PROXIMABLE_FUNCTIONALS)
-def test_functional_cuda_prox(functional: type[ProximableFunctional], parameters: Literal['scalar', 'none', 'tensor']):
+def test_functional_cuda_prox(
+    functional: type[ElementaryProximableFunctional], parameters: Literal['scalar', 'none', 'tensor']
+):
     """Test if prox and prox_convex_conj work on the GPU."""
     x = torch.tensor(1.0).cuda()
     match parameters:
@@ -393,7 +322,7 @@ def test_functional_cuda_prox(functional: type[ProximableFunctional], parameters
 
 
 @pytest.mark.parametrize('functional', FUNCTIONALS)
-def test_functional_scalar_arguments_forward(functional: type[Functional]):
+def test_functional_scalar_arguments_forward(functional: type[ElementaryFunctional]):
     """Test if the forward pass works with scalar weight and target."""
     x = torch.tensor(1.0)
     f_scalar = functional(weight=1.0, target=0.0)
@@ -406,7 +335,7 @@ def test_functional_scalar_arguments_forward(functional: type[Functional]):
 @pytest.mark.parametrize('function', ['prox', 'prox_convex_conj'])
 @pytest.mark.parametrize('functional', PROXIMABLE_FUNCTIONALS)
 def test_functional_scalar_arguments_prox(
-    functional: type[ProximableFunctional], function: Literal['prox', 'prox_convex_conj']
+    functional: type[ElementaryProximableFunctional], function: Literal['prox', 'prox_convex_conj']
 ):
     """Test if prox and prox_convex_conj work with scalar weight and target."""
     x = torch.tensor(1.0)
@@ -418,7 +347,7 @@ def test_functional_scalar_arguments_prox(
 
 
 @pytest.mark.parametrize('functional', FUNCTIONALS)
-def test_functional_default_arguments_forward(functional: type[Functional]):
+def test_functional_default_arguments_forward(functional: type[ElementaryFunctional]):
     """Test if the forward pass works with default weight and target."""
     x = torch.tensor(1.0)
     f_default = functional()
@@ -431,7 +360,7 @@ def test_functional_default_arguments_forward(functional: type[Functional]):
 @pytest.mark.parametrize('function', ['prox', 'prox_convex_conj'])
 @pytest.mark.parametrize('functional', PROXIMABLE_FUNCTIONALS)
 def test_functional_default_arguments_prox(
-    functional: type[ProximableFunctional], function: Literal['prox', 'prox_convex_conj']
+    functional: type[ElementaryProximableFunctional], function: Literal['prox', 'prox_convex_conj']
 ):
     """Test if prox and prox_convex_conj work with default weight and target."""
     x = torch.tensor(1.0)
